More refactoring. Moved transaction list under transactions view directory. Divided transaction list into Debit and Credit. Removed showing plug (+) or minus (-) to avoid confusion since in each list it can mean a different thing

+++ app/Http/Controllers/BalanceController.php
@@ -132,6 +132,8 @@ class BalanceController extends Controller
     public function destroy(Balance $balance)
     {
 
+        $this->authorize('update', $balance);
+
         //delete children relations tied to balances first then balance
         $balance->transactions()->delete();
         $balance->budget_categories()->delete();
